"Refactored size-related variables to Size type"
Replaced every instance of 'long' type for size and offset related variables to 'Size' type in the HttpDownloader. The 'Size' type introduces a more formalized way to represent sizes across the application, thereby enhancing code readability and maintainability. The 'Size' type also provides various helper methods and security checks that weren't available with the 'long' type. This allows for a more standardized and safer way to manipulate the respective variables across different classes and methods.