Changed: Disabled GC by Default [Please Read Comment]
This change disables the GC by default. It seems that due to a miscommunication, there was a misinterpretation that very likely Tim and I held in terms of how the GC should work. We were under the impression that GC should be 'I delete a loadout,

It appears that the Library is intended to be global, and instead the GC should work by mapping the Loadout.

This Commit it intended towards making it mergeable into main, as to avoid exploding the branch any further. In a future PR, the GC code will be changed such that it instead looks at the library items. That should be roughly a day's worth of work.