Changed: Lift out the non-generic parts of TreeDataGridItemModel out to the base class.
This lifts out parts of the `TreeDataGridItemModel<TModel, TKey>` to its base variant, `TreeDataGridItemModel`, while exposing more of its functionality via the interface.

I found that there are situations (e.g. in the design) that require me to play with the state of each individual row. For example; the rule 'if you're a free user and there are multiple mods to update under a single page, then expand the mod page row'.

Previously access to this functionality wasn't (really) possible without a nasty cast into a generic; I would rather avoid that, so we're going for a non-generic interface.

While going at that, I noticed that more behaviour could be lifted out to the base class, so I did, in case there are future needs similar to this. I didn't notably lift out model activation, I could; but I'm deciding to leave that be for now, since that involves a tiny bit of generic logic I want to avoid touching; since we'd kind of be breaking the nice chain of notes.