Fix GetGameVersion for Stardew Valley
The SMAPI API doesn't return any data if the `ISemanticVersion` we're
passing on is "non-standard". Although the API type allows any
`ISemanticVersion` implementation, it apparently doesn't like anything
other than versions created using the `SemanticVersion(Version)`
constructor.

This fix changes the behavior of how we create `SemanticVersion` to the
old behavior of passing the `System.Version` we get from the Stardew
Valley DLL to the `SemanticVersion` constructor.

Fixes #2705.