Remove timeout and cancel exception from CliServer
Before, the CliServer would create a cancellation token source that cancels
after 1 second inside a while loop and use that token for
`AcceptTcpClientAsync`. This meant that every second we're creating new
token sources and every second we'd throw an
`OperationCanceledException`.

This seemed very unusual, I couldn't figure out why we'd want that, so I
updated the code to just use the main cancellation token.