Fix duplicate collection downloads on upload
When uploading a collection revision, we can either create a new
revision or update an existing revision. The latter causes issues in our
data model because we previously assumed that revisions are mutable.

This "fix" is more of a hack and will remove all the previous downloads
before adding the new ones. It will throw an exception if any of the
previous download entities are referenced by anything.